Configure Javascript tests to allow running in browser
