+++ modules/swagger-codegen/src/main/resources/Javascript/package.mustache
@@ -12,6 +12,7 @@
   },
   "devDependencies": {
     "mocha": "~2.3.4",
+    "expect.js": "~0.3.1",
     "mockrequire": "~0.0.5",
     "domino": "~1.0.20",
     "xmlhttprequest": "~1.8.0"
